Break sdk-tools dependency cycle
