+++ packages/agent-sdk-ts/src/sdk/types/tools.ts
@@ -1,0 +1,15 @@
+import type { EventLog } from '../runtime/EventLog';
+import type { SecretRegistry } from '../runtime/SecretRegistry';
+import type { LocalWorkspace } from '../../workspace/LocalWorkspace';
+
+export interface ToolContext {
+  workspace: LocalWorkspace;
+  events?: EventLog;
+  secrets?: SecretRegistry;
+}
+
+export interface ToolHandler<TArgs, TResult> {
+  name: string;
+  validate(input: unknown): TArgs;
+  execute(args: TArgs, context: ToolContext): Promise<TResult>;
+}
